// src/app/enums.ts
export enum SenderType {
    User = 'user',
    Bot = 'bot',
}

export enum MessageRatingType {
    None = 'none',
    ThumbUp = 'upvote',
    ThumbDown = 'downvote',
}

export enum MessageStatusType {
    Pending = 'pending',
    Received = 'received',
    Error = 'error',
    System = 'system',
}